+++ backend/internal/model/article.go
