+++ backend/internal/model/article/linkages/to/many/query.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/go-pg/pg"
 
-	"github.com/yoshihiro-shu/draft-backend/model/article/table"
+	"github.com/yoshihiro-shu/draft-backend/internal/model/article/table"
 )
 
 type (
