+++ backend/internal/model/article/table/type.go
